backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
idna==3.10
pillow==11.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
yarg==0.1.10
